+++ types/core/struct.go
@@ -9,10 +9,6 @@ var (
 	App *Core
 )
 
-func init() {
-	App = new(Core)
-}
-
 // Core struct contains all the required fields for Statping. All application settings
 // will be saved into 1 row in the 'core' table. You can use the core.CoreApp
 // global variable to interact with the attributes to the application, such as services.